#32 [refactor] OurRulesUseCase 추가
