+++ domain/src/main/java/hous/release/domain/repository/OurRulesRepository.kt
@@ -1,8 +1,8 @@
 package hous.release.domain.repository
 
-import hous.release.domain.entity.OurRulesContent
+import hous.release.domain.entity.response.OurRule
 import kotlinx.coroutines.flow.Flow
 
 interface OurRulesRepository {
-    fun fetchOurRulesContent(): Flow<OurRulesContent>
+    fun fetchOurRulesContent(): Flow<List<OurRule>>
 }
